requirements.txt
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.13.0
urllib3==1.25.7
python-binance
pandas==1.1.5
pandas_ta
numpy==1.19.5
python-dateutil==2.8.1
pytz==2021.1
psycopg2-binary